Override Elements reset on accordion <h2> height
https://github.com/alphagov/govuk_publishing_components/pull/1386 reports
that height of accordion sections on mobile don't match GOV.UK Design System.
This is because of the inclusion of GOV.UK Elements which resets `<h2>` to
inherit font-size and line-height from `<body>`.

To fix, set the font-size and line-height match 24px/1.2 if Elements is
detected. Note that `govuk-font` mixin can't be used here as it uses the
responsive font size map - having a different font size on mobile would
reduce the height there which wouldn't match the current behaviour.

Note that as we don't set font-size and line-height explicitly on the `<h2>`,
they are dependent on the user agent stylesheet (unless something like Elements
overrides it). For this reason we could consider setting 24px/1.2 on accordion
<h2> elements without checking for Elements. However we don't currently have
evidence that this is causing users problems and it seems somewhat unlikely that
it would.